Added BaseEntity abstract class and made changes to Email and Site entities
