+++ src/main/java/com/emailspringproject/emailholder/domain/entities/BaseEntity.java
@@ -1,0 +1,19 @@
+package com.emailspringproject.emailholder.domain.entities;
+
+import jakarta.persistence.*;
+
+@MappedSuperclass
+public abstract class BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+}
